+++ src/main/java/com/calculadora/services/CalculadoraService.java
@@ -4,6 +4,6 @@ import java.math.BigDecimal;
 
 public interface CalculadoraService {
 
-	BigDecimal ejecutar(String operador, BigDecimal ope1, BigDecimal ope2) throws Exception;
+	Double ejecutar(String operador, BigDecimal ope1, BigDecimal ope2) throws Exception;
 
 }
